Upload Lab Assignment 4: Event Handling & State Lifting
Practice state lifting, event handling, and form interaction by implementing the ability to add tasks to the ToDo List using a form. You will create a parent component to manage the state of tasks, pass down the necessary functions to child components, and update the list of tasks based on user input.